Rename module display -> forme
